pyproject.toml
[project]
name = "bilibili-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bilibili-api-python>=17.1.4",
    "mcp[cli]>=1.5.0",
]
